+++ CHANGELOG.md
@@ -10,6 +10,13 @@ The format is based on [Keep a Changelog], and this project adheres to
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
+## [Unreleased]
+
+### Fixed
+
+- Fixed incorrect panic value when calling
+  `ProjectionConfigurer.DeliveryPolicy()` with a `nil` policy.
+
 ## [0.13.0] - 2024-03-26
 
 - **[BC]** Updated to Dogma v0.13.0